refactor: move DeviceRegistry to utils.py and simplify parameter conversion logic
